pyproject.toml
[project]
name = "deepfakes-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "torch",
    "torchvision",
    "transformers",
    "pillow",
    "numpy",
    "safetensors",
    "pyqt6",
]
